requirements.txt
amqp==5.2.0
apturl==0.5.2
async-timeout==4.0.3
attrs==21.2.0
bcrypt==3.2.0
billiard==4.2.0
blinker==1.7.0
Brlapi==0.8.3
celery==5.3.6
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
fasteners==0.14.1
filelock==3.9.0
Flask==3.0.0
Flask-WTF==1.2.1
future==0.18.2
greenlet==1.1.2
gyp==0.1
httplib2==0.20.2
idna==3.3
importlib-metadata==6.8.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==3.2.0
keyring==23.5.0
kombu==5.3.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.3
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.3
netifaces==0.11.0
numpy==1.26.1
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
pexpect==4.8.0
pika==1.3.2
Pillow==10.1.0
platformdirs==2.6.2
prompt-toolkit==3.0.43
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynvim==0.4.2
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==0.19.2
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
redis==5.0.1
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
six==1.16.0
ssh-import-id==5.11
systemd-python==234
texttable==1.6.4
typing_extensions==4.9.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uWSGI==2.0.21
vine==5.1.0
virtualenv==20.17.1
wadllib==1.3.6
wcwidth==0.2.12
websocket-client==1.2.3
Werkzeug==3.0.1
WTForms==3.1.1
xdg==5
xkit==0.0.0
zipp==3.17.0
